# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mailhub management API",
    "version": "2.7.0",
    "description": "Active single-owner management REST API. Bearer credentials can be static API keys or configured OAuth access tokens. x-required-permission records the exact scope; compliance also enforces configured domain/action grants. Errors are generally plain text. Legacy handlers may accept undocumented method aliases; use only the documented methods. JMAP is a separate contract."
  },
  "servers": [
    {
      "url": "https://{host}",
      "variables": {
        "host": {
          "default": "localhost:8080",
          "description": "Configured api.rest_addr behind the deployment TLS endpoint."
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/queue/stats": {
      "get": {
        "operationId": "get_api_v1_queue_stats",
        "summary": "Queue and storage counters",
        "responses": {
          "200": {
            "description": "Queue and storage counters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueStats"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "queue:read"
      },
      "head": {
        "operationId": "head_api_v1_queue_stats",
        "summary": "Queue and storage counters (headers only)",
        "responses": {
          "200": {
            "description": "Queue and storage counters"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "queue:read"
      }
    },
    "/api/v1/queue/pending": {
      "get": {
        "operationId": "get_api_v1_queue_pending",
        "summary": "Pending messages",
        "responses": {
          "200": {
            "description": "Pending messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalEntry"
                  },
                  "nullable": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "queue:read",
        "parameters": [
          {
            "name": "tier",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Omit for all tiers."
          }
        ]
      },
      "head": {
        "operationId": "head_api_v1_queue_pending",
        "summary": "Pending messages (headers only)",
        "responses": {
          "200": {
            "description": "Pending messages"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "queue:read",
        "parameters": [
          {
            "name": "tier",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Omit for all tiers."
          }
        ]
      }
    },
    "/api/v1/dlq/list": {
      "get": {
        "operationId": "get_api_v1_dlq_list",
        "summary": "Failed messages",
        "responses": {
          "200": {
            "description": "Failed messages",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalEntry"
                  },
                  "nullable": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "queue:read"
      },
      "head": {
        "operationId": "head_api_v1_dlq_list",
        "summary": "Failed messages (headers only)",
        "responses": {
          "200": {
            "description": "Failed messages"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "queue:read"
      }
    },
...
    "/api/v1/mailboxes": {
      "get": {
        "operationId": "get_api_v1_mailboxes",
        "summary": "List mail accounts",
        "responses": {
          "200": {
            "description": "List mail accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mailbox"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "501": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "mailboxes:read"
      },
      "post": {
        "operationId": "post_api_v1_mailboxes",
        "summary": "Create mail account",
        "responses": {
          "201": {
            "description": "Create mail account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mailbox"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "501": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "mailboxes:write",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MailboxCreate"
              }
            }
          }
        },
        "x-max-body-bytes": 65536
      }
    },
...
    "/api/v1/policies": {
      "get": {
        "operationId": "get_api_v1_policies",
        "summary": "List policies",
        "responses": {
          "200": {
            "description": "List policies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "policies:read"
      },
      "post": {
        "operationId": "post_api_v1_policies",
        "summary": "Create policy",
        "responses": {
          "201": {
            "description": "Create policy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicySaved"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "policies:write",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Policy"
              }
            }
          }
        },
        "x-max-body-bytes": 2097152
      }
    },
...
    "/api/v1/policies/stats": {
      "get": {
        "operationId": "get_api_v1_policies_stats",
        "summary": "Policy engine counters",
        "responses": {
          "200": {
            "description": "Policy engine counters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyStats"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "policies:read"
      }
    },
    "/api/v1/policies/reload": {
      "post": {
        "operationId": "post_api_v1_policies_reload",
        "summary": "Reload managed policies",
        "responses": {
          "200": {
            "description": "Reload managed policies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyReloaded"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "501": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "policies:write"
      }
    },
    "/api/v1/mailstorm": {
      "get": {
        "operationId": "get_api_v1_mailstorm",
        "summary": "Mailstorm guard status",
        "responses": {
          "200": {
            "description": "Mailstorm guard status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mailstorm"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "mailstorm:read"
      }
    },
...
    "/api/v1/quarantine": {
      "get": {
        "operationId": "get_api_v1_quarantine",
        "summary": "List held message metadata",
        "responses": {
          "200": {
            "description": "List held message metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Quarantined"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "quarantine:read"
      }
    },
...
    "/api/v1/bounce/config": {
      "get": {
        "operationId": "get_api_v1_bounce_config",
        "summary": "Effective bounce configuration",
        "responses": {
          "200": {
            "description": "Effective bounce configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BounceConfig"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "bounce:read"
      }
    },
    "/api/v1/bounce/reports": {
      "get": {
        "operationId": "get_api_v1_bounce_reports",
        "summary": "Tracked incoming DSNs",
        "responses": {
          "200": {
            "description": "Tracked incoming DSNs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalEntry"
                  },
                  "nullable": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "405": {
            "$ref": "#/components/responses/Error"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "x-required-permission": "bounce:read"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Static API key or configured OAuth access token. Mailbox passwords/Basic Auth are not management credentials."
      }
    },
    "schemas": {
...
      "Mailbox": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "username",
          "email",
          "enabled"
        ],
        "additionalProperties": false
      },
...
      "PolicyList": {
        "type": "object",
        "properties": {
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Policy"
            },
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "policies",
          "count"
        ]
      },
...
    },
    "responses": {
      "Error": {
        "description": "Plain-text failure. Inspect HTTP status before decoding. Authorization failures may be 401 or 403 depending on credential mode.",
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}